+++ emmet-builders/emmet/builders/materials/summary.py
@@ -14,6 +14,7 @@ class SummaryBuilder(Builder):
         materials,
         thermo,
         xas,
+        chemenv,
         grain_boundaries,
         electronic_structure,
         magnetism,
@@ -37,6 +38,7 @@ class SummaryBuilder(Builder):
         self.materials = materials
         self.thermo = thermo
         self.xas = xas
+        self.chemenv = chemenv
         self.grain_boundaries = grain_boundaries
         self.electronic_structure = electronic_structure
         self.magnetism = magnetism
@@ -61,6 +63,7 @@ class SummaryBuilder(Builder):
                 materials,
                 thermo,
                 xas,
+                chemenv,
                 grain_boundaries,
                 electronic_structure,
                 magnetism,
@@ -118,6 +121,9 @@ class SummaryBuilder(Builder):
                 HasProps.materials.value: materials_doc,
                 HasProps.thermo.value: self.thermo.query_one({self.thermo.key: entry}),
                 HasProps.xas.value: list(self.xas.query({self.xas.key: entry})),
+                HasProps.chemenv.value: self.chemenv.query_one(
+                    {self.chemenv.key: entry}
+                ),
                 HasProps.grain_boundaries.value: list(
                     self.grain_boundaries.query({self.grain_boundaries.key: entry})
                 ),
@@ -141,7 +147,8 @@ class SummaryBuilder(Builder):
                 ),
                 HasProps.insertion_electrodes.value: list(
                     self.insertion_electrodes.query(
-                        {"material_ids": entry}, [self.insertion_electrodes.key],
+                        {"material_ids": entry},
+                        [self.insertion_electrodes.key],
                     )
                 ),
                 HasProps.surface_properties.value: self.surfaces.query_one(